bai6

+++ bai6/mongodb_spring/src/main/java/com/example/mongodb_spring/Service/Implement/BookImpl.java
@@ -1,11 +1,15 @@
 package com.example.mongodb_spring.Service.Implement;
 
+import com.example.mongodb_spring.Exceptions.ApiException;
 import com.example.mongodb_spring.Model.Book;
 import com.example.mongodb_spring.Service.BookService;
 import com.example.mongodb_spring.repo.BookRepo;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -18,32 +22,34 @@ public class BookImpl implements BookService {
 
     @Autowired
     private BookRepo bookRepo;
-
-    @Override
-    public List<Book> getAll() {
-        return bookRepo.findAll();
-    }
-
     @Override
     public Book findBookById(String id) {
         Optional<Book> bookOp = this.bookRepo.findById(id);
-        return bookOp.orElse(null);
+        return bookOp.orElse(new Book());
     }
 
     @Override
     public String addBook(Book book) {
+        if (book.getName() == null || book.getAuthor() == null || book.getPublicationDate() == null){
+            throw new ApiException("Can not Empty");
+        }
         this.bookRepo.save(book);
         return "Done!";
     }
-
     @Override
     public String deleteBook(String id) {
+        if (this.findBookById(id).equals(new Book())){
+            throw  new ApiException("Book Not Exist");
+        }
         this.bookRepo.deleteById(id);
         return "Delete success";
     }
 
     @Override
     public String updateBook(String id, Book book) {
+        if (book.getName() == null || book.getAuthor() == null || book.getPublicationDate() == null){
+            throw new ApiException("Can not Empty");
+        }
         Book bookDB = this.findBookById(id);
         bookDB.setName(book.getName());
         bookDB.setAuthor(book.getAuthor());
@@ -53,16 +59,30 @@ public class BookImpl implements BookService {
         return "Done!";
     }
 
-    @Override
-    public List<Book> findByAuthorAndName(String author, String name) {
-        return this.bookRepo.findBookByAuthorContainsAndNameContains(author,name);
-
-    }
 
     @Override
     public List<Book> findBookByDate(Date start, Date end) {
-        return this.bookRepo.findBookByPublicationDateBetween(start,end);
+        Pageable pageable = PageRequest.of(0,5, Sort.by("publicationDate").descending());
+        Page<Book> pageResult = bookRepo.findBookByPublicationDateBetween(start,end, pageable);
+        return pageResult.toList();
     }
 
+    @Override
+    public List<Book> findByText(String author) {
+        return this.bookRepo.findByText(author);
+    }
 
+    @Override
+    public List<Book> getAllBook(int page, int size) {
+        if (page < 0){
+            throw new ApiException("page >= 0");
+        } else if (size < 1) {
+            throw new ApiException("size > 1");
+        }
+        else {
+            Pageable paging = PageRequest.of(page,size, Sort.by("publicationDate").descending());
+            Page<Book> pageResult = bookRepo.findAll(paging);
+            return pageResult.toList();
+        }
+    }
 }
